+++ test/csv.js
@@ -1,0 +1,111 @@
+const assert = require("assert");
+const ripeCommons = require("..");
+
+describe("CSV", () => {
+    describe("#parseCsv", () => {
+        it("should allow simple CSV parsing", () => {
+            const result = ripeCommons.parseCsv(
+                "name,title\nJoão Magalhães,mr\nGabriel Candal,mr\n"
+            );
+            assert.deepStrictEqual(result, [
+                ["name", "title"],
+                ["João Magalhães", "mr"],
+                ["Gabriel Candal", "mr"]
+            ]);
+        });
+
+        it("should allow simple CSV object parsing", () => {
+            const result = ripeCommons.parseCsv(
+                "name,title\nJoão Magalhães,mr\nGabriel Candal,mr\n",
+                true
+            );
+            assert.deepStrictEqual(result, [
+                { name: "João Magalhães", title: "mr" },
+                { name: "Gabriel Candal", title: "mr" }
+            ]);
+        });
+
+        it("should handle extra elements in line", () => {
+            let result;
+
+            result = ripeCommons.parseCsv(
+                "name,title\nJoão Magalhães,mr \nGabriel Candal,mr   \n",
+                true
+            );
+            assert.deepStrictEqual(result, [
+                { name: "João Magalhães", title: "mr" },
+                { name: "Gabriel Candal", title: "mr" }
+            ]);
+
+            result = ripeCommons.parseCsv(
+                "name,title\nJoão Magalhães,mr \nGabriel Candal,mr   \n",
+                true,
+                false
+            );
+            assert.deepStrictEqual(result, [
+                { name: "João Magalhães", title: "mr " },
+                { name: "Gabriel Candal", title: "mr   " },
+                { name: "", title: undefined }
+            ]);
+        });
+    });
+
+    describe("#_parseCsvComplex", () => {
+        it("should allow simple CSV parsing", () => {
+            const result = ripeCommons._parseCsvComplex(
+                "name,title\nJoão Magalhães,mr\nGabriel Candal,mr\n"
+            );
+            assert.deepStrictEqual(result, [
+                ["name", "title"],
+                ["João Magalhães", "mr"],
+                ["Gabriel Candal", "mr"]
+            ]);
+        });
+
+        it("should allow simple CSV object parsing", () => {
+            const result = ripeCommons._parseCsvComplex(
+                "name,title\nJoão Magalhães,mr\nGabriel Candal,mr\n",
+                true
+            );
+            assert.deepStrictEqual(result, [
+                { name: "João Magalhães", title: "mr" },
+                { name: "Gabriel Candal", title: "mr" }
+            ]);
+        });
+
+        it("should allow complex CSV object parsing", () => {
+            const result = ripeCommons._parseCsvComplex(
+                'name,title\n"João,{Maga""lhães}",mr\nGabriel Candal,mr\n',
+                true
+            );
+            assert.deepStrictEqual(result, [
+                { name: 'João,{Maga"lhães}', title: "mr" },
+                { name: "Gabriel Candal", title: "mr" }
+            ]);
+        });
+
+        it("should handle extra elements in line", () => {
+            let result;
+
+            result = ripeCommons._parseCsvComplex(
+                "name,title\nJoão Magalhães,mr \nGabriel Candal,mr   \n",
+                true
+            );
+            assert.deepStrictEqual(result, [
+                { name: "João Magalhães", title: "mr" },
+                { name: "Gabriel Candal", title: "mr" }
+            ]);
+
+            result = ripeCommons._parseCsvComplex(
+                "name,title\nJoão Magalhães,mr \nGabriel Candal,mr   \n",
+                true,
+                false
+            );
+            assert.deepStrictEqual(result, [
+                { name: "João Magalhães", title: "mr " },
+                { name: "Gabriel Candal", title: "mr   " },
+                { name: "", title: undefined }
+            ]);
+        });
+    });
+});