+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* Refactored the ACL function out of the fastify module to be used by express aswell - [ripe-orchestra/#16](https://github.com/ripe-tech/ripe-orchestra/issues/16)
+* Refactored the ACL function out of the fastify module to be used by express as well - [ripe-orchestra/#16](https://github.com/ripe-tech/ripe-orchestra/issues/16)
 
 ### Fixed