+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+*
+
+### Fixed
+
+*
+
+## [0.9.0] - 2022-04-24
+
+### Added
+
 * Extracted CSV functions from `ripe-util-vue` - [ripe-pulse/#302](https://github.com/ripe-tech/ripe-pulse/issues/302)
 * Express middleware not found and error handlers - [ripe-orchestra/#16](https://github.com/ripe-tech/ripe-orchestra/issues/16)
 * Support for more complex building and parsing of CSV data and files
@@ -16,10 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Refactored the ACL function out of the fastify module to be used by express as well - [ripe-orchestra/#16](https://github.com/ripe-tech/ripe-orchestra/issues/16)
-
-### Fixed
-
-*
 
 ## [0.8.1] - 2022-02-07